refactor: redesign SSO panels with collapsible config, provider logos, and inline enable toggle
